+++ config.lua
@@ -89,7 +89,7 @@ local function standingColorsSet(value)
     end
 
     local REP_COLORS = copy(Addon.CONST.REP_COLORS)
-    Debug:Table(REP_COLORS, "colors")
+    Debug:Table("colors", REP_COLORS)
     if value == "blizzard" then Addon.db.profile.Colors = REP_COLORS.blizzardColors end
     if value == "ascii" then Addon.db.profile.Colors = REP_COLORS.asciiColors end
     if value == "wowpro" then Addon.db.profile.Colors = REP_COLORS.wowproColors end